Relocate appModule to com.github.kr328.clash.app.di package
